# public/swagger.json
{
	"components": {
...
		"schemas": {
			"PlayerLevel": {
				"enum": [
					"NEWBIE",
					"BEGINNER",
					"INTERMEDIATE",
					"ADVANCED",
					"PRO"
				],
				"type": "string"
			},
			"RegisterRequest": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"level": {
						"$ref": "#/components/schemas/PlayerLevel"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"password",
					"nickname",
					"level",
					"address",
					"city",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlayersResponse": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"level": {
						"$ref": "#/components/schemas/PlayerLevel"
					},
					"address": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"nickname",
					"level",
					"address",
					"email",
					"city",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AppError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILoginRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePlayerRequest": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"level": {
						"$ref": "#/components/schemas/PlayerLevel"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"RacketLevels": {
				"enum": [
					"JUNIOR",
					"RECREATIONAL",
					"PROFESSIONAL"
				],
				"type": "string"
			},
			"RacketResponse": {
				"properties": {
					"brand": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"level": {
						"$ref": "#/components/schemas/RacketLevels"
					},
					"headSizeInch": {
						"type": "number",
						"format": "double"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"stringPattern": {
						"type": "string"
					},
					"recommendedStrings": {
						"type": "string"
					}
				},
				"required": [
					"brand",
					"model",
					"year",
					"weight",
					"level",
					"headSizeInch",
					"balance",
					"stringPattern",
					"recommendedStrings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateRacketRequest": {
				"properties": {
					"brand": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"level": {
						"$ref": "#/components/schemas/RacketLevels"
					},
					"headSizeInch": {
						"type": "number",
						"format": "double"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"stringPattern": {
						"type": "string"
					},
					"recommendedStrings": {
						"type": "string"
					}
				},
				"required": [
					"brand",
					"model",
					"year",
					"weight",
					"level",
					"headSizeInch",
					"balance",
					"stringPattern",
					"recommendedStrings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssignRacketRequest": {
				"properties": {
					"playerEid": {
						"type": "string"
					},
					"racketEid": {
						"type": "string"
					}
				},
				"required": [
					"playerEid",
					"racketEid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRacketRequest": {
				"properties": {
					"brand": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"level": {
						"$ref": "#/components/schemas/RacketLevels"
					},
					"headSizeInch": {
						"type": "number",
						"format": "double"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"stringPattern": {
						"type": "string"
					},
					"recommendedStrings": {
						"type": "string"
					}
				},
				"required": [
					"brand",
					"model",
					"year",
					"weight",
					"level",
					"headSizeInch",
					"balance",
					"stringPattern",
					"recommendedStrings"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BookingResponse": {
				"properties": {
					"entityId": {
						"type": "string"
					},
					"clubName": {
						"type": "string"
					},
					"clubAddress": {
						"type": "string"
					},
					"clubCity": {
						"type": "string"
					},
					"courtName": {
						"type": "string"
					},
					"courtSurface": {
						"type": "string"
					},
					"from": {
						"type": "number",
						"format": "double"
					},
					"to": {
						"type": "number",
						"format": "double"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"playerFirstName": {
						"type": "string"
					},
					"playerLastName": {
						"type": "string"
					},
					"bookingType": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"entityId",
					"clubName",
					"clubAddress",
					"clubCity",
					"courtName",
					"courtSurface",
					"from",
					"to",
					"totalPrice",
					"playerFirstName",
					"playerLastName",
					"bookingType",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBookingRequest": {
				"properties": {
					"court": {
						"type": "string"
					},
					"from": {
						"type": "number",
						"format": "double"
					},
					"to": {
						"type": "number",
						"format": "double"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"court",
					"from",
					"to",
					"totalPrice",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MatchResponse": {
				"properties": {
					"entityId": {
						"type": "string"
					},
					"firstTeam": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"secondTeam": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"result": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"clubName": {
						"type": "string"
					},
					"courtName": {
						"type": "string"
					},
					"courtSurface": {
						"type": "string"
					}
				},
				"required": [
					"entityId",
					"firstTeam",
					"secondTeam",
					"result",
					"date",
					"clubName",
					"courtName",
					"courtSurface"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tennis",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/players/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequest"
							}
						}
					}
				}
			}
		},
		"/players/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/PlayersResponse"
										},
										{}
									]
								}
							}
						}
					},
					"400": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AppError"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ILoginRequest"
							}
						}
					}
				}
			}
		},
		"/players": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PlayersResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/players/{entityId}": {
			"get": {
				"operationId": "GetByEntityId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayersResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeletePlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdatePlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePlayerRequest"
							}
						}
					}
				}
			}
		},
		"/players/city/{city}": {
			"get": {
				"operationId": "GetPlayersByCity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PlayersResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "city",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/players/level/{level}": {
			"get": {
				"operationId": "GetPlayersByLevel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PlayersResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "level",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/PlayerLevel"
						}
					}
				]
			}
		},
...
		"/rackets": {
			"get": {
				"operationId": "GetRackets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RacketResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Rackets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateRacket",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Rackets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRacketRequest"
							}
						}
					}
				}
			}
		},
		"/rackets/all": {
			"get": {
				"operationId": "GetAllRackets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RacketResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Rackets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/rackets/assign": {
			"post": {
				"operationId": "AssignRacketToPlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Rackets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignRacketRequest"
							}
						}
					}
				}
			}
		},
		"/rackets/{entityId}": {
			"delete": {
				"operationId": "DeleteRacket",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Rackets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetRacket",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RacketResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Rackets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateRacket",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Rackets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRacketRequest"
							}
						}
					}
				}
			}
		},
...
		"/bookings/{entityId}": {
			"delete": {
				"operationId": "DeleteBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBookingRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetByEntityId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/matches/player/{entityId}": {
			"get": {
				"operationId": "GetPlayersMatches",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MatchResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Matches"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}